+++ cli/amazeeio_cachalot/check_env.rb
@@ -12,7 +12,7 @@ class CheckEnv
       puts "\e[32mYour environment variables are already set correctly.\e[0m"
     else
       puts "\e[33mTo connect the Docker client to the Docker daemon, please run:\e[0m"
-      puts "eval $(amazeeio-cachalot shellinit)"
+      puts "eval $(amazeeio-cachalot env)"
     end
   end
 
@@ -28,7 +28,7 @@ class CheckEnv
   def print
     expected.each { |name,value| puts "export #{name}=#{value}" }
     puts "# run this command, to setup your docker:"
-    puts "# eval $(amazeeio-cachalot shellinit)"
+    puts "# eval $(amazeeio-cachalot env)"
   end
 
   def set?